refactor(useOnLongPress): Split desktop / mobile behaviors
